requirements.txt
streamlit>=1.31.0
openai
pinecone-client
python-dotenv
numpy
scikit-learn
python-dateutil
pymongo==4.6.1 